+++ core/src/db/meta/MetaTypes.h
@@ -25,7 +25,6 @@ namespace engine {
 namespace meta {
 
 constexpr int32_t DEFAULT_ENGINE_TYPE = (int)EngineType::FAISS_IDMAP;
-constexpr int32_t DEFAULT_NLIST = 16384;
 constexpr int32_t DEFAULT_METRIC_TYPE = (int)MetricType::L2;
 constexpr int32_t DEFAULT_INDEX_FILE_SIZE = GB;
 constexpr char CURRENT_VERSION[] = MILVUS_VERSION;