+++ core/src/index/thirdparty/faiss/FaissHook.h
@@ -6,15 +6,17 @@
 #include <vector>
 #include <stddef.h>
 #include <string>
+#include <faiss/impl/ScalarQuantizer.h>
 #include <faiss/impl/ScalarQuantizerOp.h>
+#include <faiss/MetricType.h>
 
 namespace faiss {
 
 typedef float (*fvec_func_ptr)(const float*, const float*, size_t);
 
-typedef SQDistanceComputer* (*sq_get_func_ptr)(QuantizerType, size_t, const std::vector<float>&);
-typedef Quantizer* (*sq_sel_func_ptr)(QuantizerType, size_t, const std::vector<float>&);
-
+typedef SQDistanceComputer* (*sq_get_distance_computer_func_ptr)(MetricType, QuantizerType, size_t, const std::vector<float>&);
+typedef Quantizer* (*sq_sel_quantizer_func_ptr)(QuantizerType, size_t, const std::vector<float>&);
+typedef InvertedListScanner* (*sq_sel_inv_list_scanner_func_ptr)(MetricType, const ScalarQuantizer*, const Index*, size_t, bool, bool);
 
 extern bool faiss_use_avx512;
 extern bool faiss_use_avx2;
@@ -25,9 +27,9 @@ extern fvec_func_ptr fvec_L2sqr;
 extern fvec_func_ptr fvec_L1;
 extern fvec_func_ptr fvec_Linf;
 
-extern sq_get_func_ptr sq_get_distance_computer_L2;
-extern sq_get_func_ptr sq_get_distance_computer_IP;
-extern sq_sel_func_ptr sq_sel_quantizer;
+extern sq_get_distance_computer_func_ptr sq_get_distance_computer;
+extern sq_sel_quantizer_func_ptr sq_sel_quantizer;
+extern sq_sel_inv_list_scanner_func_ptr sq_sel_inv_list_scanner;
 
 extern bool support_avx512();
 extern bool support_avx2();