+++ core/unittest/db/test_hybrid_db.cpp
@@ -32,7 +32,7 @@ namespace {
 static const char* COLLECTION_NAME = "test_hybrid";
 static constexpr int64_t COLLECTION_DIM = 128;
 static constexpr int64_t SECONDS_EACH_HOUR = 3600;
-static constexpr int64_t FIELD_NUM = 4;
+static constexpr int64_t FIELD_NUM = 3;
 static constexpr int64_t NQ = 10;
 static constexpr int64_t TOPK = 10;
 
@@ -52,12 +52,19 @@ BuildCollectionSchema(milvus::engine::meta::CollectionSchema& collection_schema,
     fields[0].field_type_ = (int)milvus::engine::meta::hybrid::DataType::INT32;
     fields[1].field_type_ = (int)milvus::engine::meta::hybrid::DataType::INT64;
     fields[2].field_type_ = (int)milvus::engine::meta::hybrid::DataType::FLOAT;
-    fields[3].field_type_ = (int)milvus::engine::meta::hybrid::DataType::VECTOR;
-    fields_schema.fields_schema_ = fields;
 
-    attr_type.insert(std::make_pair("field_0", milvus::engine::meta::hybrid::DataType::INT32));
-    attr_type.insert(std::make_pair("field_1", milvus::engine::meta::hybrid::DataType::INT64));
-    attr_type.insert(std::make_pair("field_2", milvus::engine::meta::hybrid::DataType::FLOAT));
+    for (uint64_t i = 0; i < FIELD_NUM; ++i) {
+        attr_type.insert(
+            std::make_pair(fields[i].field_name_, (milvus::engine::meta::hybrid::DataType)fields[i].field_type_));
+    }
+
+    milvus::engine::meta::hybrid::FieldSchema schema;
+    schema.field_name_ = "field_3";
+    schema.collection_id_ = COLLECTION_NAME;
+    schema.field_type_ = (int)(milvus::engine::meta::hybrid::DataType::VECTOR);
+    fields.emplace_back(schema);
+
+    fields_schema.fields_schema_ = fields;
 }
 
 void
@@ -174,6 +181,7 @@ ConstructGeneralQuery(milvus::query::GeneralQueryPtr& general_query, milvus::que
     query_ptr->root = general_query->bin;
     query_ptr->vectors.insert(std::make_pair(vector_placeholder, vector_query));
 }
+
 }  // namespace
 
 TEST_F(DBTest, HYBRID_DB_TEST) {