+++ core/unittest/db/test_meta.cpp
@@ -56,7 +56,7 @@ TEST_F(MetaTest, COLLECTION_TEST) {
     ASSERT_TRUE(status.ok());
 }
 
-TEST_F(MetaTest, FALID_TEST) {
+TEST_F(MetaTest, FAILED_TEST) {
     fiu_init(0);
     auto options = GetOptions();
     auto collection_id = "meta_test_table";
@@ -66,7 +66,7 @@ TEST_F(MetaTest, FALID_TEST) {
 
     {
         FIU_ENABLE_FIU("SqliteMetaImpl.ValidateMetaSchema.NullConnection");
-        milvus::engine::meta::SqliteMetaImpl impl(options.meta_);
+        ASSERT_ANY_THROW(milvus::engine::meta::SqliteMetaImpl impl(options.meta_));
         fiu_disable("SqliteMetaImpl.ValidateMetaSchema.NullConnection");
     }
     {
@@ -294,6 +294,20 @@ TEST_F(MetaTest, FALID_TEST) {
         status = impl_->FilesByType(collection_id, file_types, files_holder);
         ASSERT_EQ(status.code(), milvus::DB_META_TRANSACTION_FAILED);
         fiu_disable("SqliteMetaImpl.FilesByType.throw_exception");
+    }
+    {
+        milvus::engine::meta::FilesHolder files_holder;
+        std::vector<milvus::engine::meta::CollectionSchema> collection_array;
+        milvus::engine::meta::CollectionSchema schema;
+        schema.collection_id_ = collection_id;
+        collection_array.emplace_back(schema);
+        std::vector<int> file_types;
+        file_types.push_back(milvus::engine::meta::SegmentSchema::INDEX);
+        FIU_ENABLE_FIU("SqliteMetaImpl.FilesByTypeEx.throw_exception");
+        status = impl_->FilesByTypeEx(collection_array, file_types, files_holder);
+        ASSERT_EQ(status.code(), milvus::DB_META_TRANSACTION_FAILED);
+        fiu_disable("SqliteMetaImpl.FilesByTypeEx.throw_exception");
+        status = impl_->FilesByTypeEx(collection_array, file_types, files_holder);
     }
     {
         uint64_t size = 0;
@@ -566,6 +580,9 @@ TEST_F(MetaTest, ARCHIVE_TEST_DISK) {
         }
         ++i;
     }
+
+    status = impl.GetCollectionFilesBySegmentId(table_file.segment_id_, files_holder);
+    ASSERT_TRUE(status.ok());
 
     impl.DropAll();
 }