+++ CHANGELOG.md
@@ -5,10 +5,11 @@ Please mark all change in change log and use the issue from GitHub
 # Milvus 0.11.0 (TBD)
 
 ## Bug
-\#2487 Remove timeout when creating collection in dev test
+-   \#2487 Remove timeout when creating collection in dev test
+-   \#2551 Fix test_hybrid_db and test_rpc error
 
 ## Feature
-- \#2319 Redo metadata to support MVCC
+-   \#2319 Redo metadata to support MVCC
 
 ## Improvement
 
@@ -18,6 +19,8 @@ Please mark all change in change log and use the issue from GitHub
 
 ## Bug
 -   \#2367 Fix inconsistent reading and writing when using mishards
+-   \#2368 Make read node detect delete behavior
+-   \#2373 Build index for small segment waste time on waiting background index thread finish
 -   \#2394 Drop collection timeout if too many partitions created on collection
 
 ## Feature
@@ -25,8 +28,12 @@ Please mark all change in change log and use the issue from GitHub
 
 ## Improvement
 -   \#2381 Upgrade FAISS to 1.6.3
+-   \#2429 Fix Milvus 0.9.1 performance degrade issue
 -   \#2441 Improve Knowhere code coverage
 -   \#2466 optimize k-selection implemention of faiss gpu version
+-   \#2489 Add exception throw on mysql meta error
+-   \#2495 Add creating lock file failure reason.
+-   \#2516 Improve unit test coverage
 
 ## Task